==> notebooks/analysis/14_edgeR-scVSbulk.ipynb <==
### Analysis of Differentially Expressed Genes between bulk and single-cell datasets
Discover cell types that might be missing in single-cell datasets.

library(edgeR)

main_dir <- "/projects/robson-lab/research/endometriosis/"
sample_id <- "Endometriosis-revision-3"
data_dir <- "DEG/edgeR-input"
setwd(main_dir)

getwd()

**Load Data**
<br>load matrices containing cpm counts from RNAseq experiments and associated metadata

sc_matrix <- read.delim(paste("data/bulkRNAseq/",sample_id,"-sc_pseudo-bulk_CPM.txt", sep=""), sep="\t")
bs_matrix <- read.delim(paste("data/bulkRNAseq/bulk_cpm_20210719.txt", sep=""), sep="\t")
metadata <- read.delim(paste("data/bulkRNAseq/",sample_id,"-combined_metadata.txt", sep=""))

dim(sc_matrix)
dim(bs_matrix)

#### DEG on Eutopic tissue vs sc-pseudo-bulk data
**Create Eutopic Matrix**

Eutopic <- metadata[metadata$sample_type %in% c("EuE","EuC"),]
Eutopic[,"X"] <- sapply(Eutopic[,"X"], as.character)

Eutopic

sc_Eut <- sc_matrix[,(colnames(sc_matrix)) %in% c("index", Eutopic$X)]
bs_Eut <- bs_matrix[,(colnames(bs_matrix)) %in% c("gene_name", Eutopic$X)]
eut_matrix <- merge(x = sc_Eut, y = bs_Eut, by.x = "index", by.y="gene_name")

dim(eut_matrix)

**Run EdgeR**

groups <- factor(Eutopic$method)

groups

cds <- DGEList(counts = as.matrix(eut_matrix[(2:25)]), group = groups,genes = eut_matrix$index)
nc <- cpm(cds, normalized.lib.sizes=FALSE)
cds <- estimateCommonDisp(cds)
cds <- estimateTagwiseDisp(cds)
de.tgw <- exactTest(cds, pair=c("single_cell","bulk_sequencing"))
diff.dat <- topTags(de.tgw,n=nrow(cds$counts))
res <- diff.dat$table
file <- paste("analysis/",sample_id,"/bulkRNAseq/DEG-Eutopic-SCvsBS_20220106.csv",sep="")
#file2 <- paste("analysis/",sample_id,"/bulkRNAseq/DEG-Top50-SCvsBS.csv",sep="")
write.csv(res[res$FDR < 0.001,],file=file)
#write.csv((row.names(res))[0:49],file=file2)

---

#### DEG on Ectopic Peritoneal tissue vs sc-pseudo-bulk data
**Create Ectopic Matrix**

Ectopic <- metadata[metadata$sample_type %in% c("EcP","EcPA"),]
Ectopic[,"X"] <- sapply(Ectopic[,"X"], as.character)

sc_Ect <- sc_matrix[,(colnames(sc_matrix)) %in% c("index", Ectopic$X)]
bs_Ect <- bs_matrix[,(colnames(bs_matrix)) %in% c("gene_name", Ectopic$X)]
ect_matrix <- merge(x = sc_Ect, y = bs_Ect, by.x = "index", by.y="gene_name")

dim(ect_matrix)

**Run EdgeR**

groups <- factor(Ectopic$method)
cds <- DGEList(counts = as.matrix(ect_matrix[(2:22)]), group = groups,genes = ect_matrix$index)
nc <- cpm(cds, normalized.lib.sizes=FALSE)
cds <- estimateCommonDisp(cds)
cds <- estimateTagwiseDisp(cds)
de.tgw <- exactTest(cds, pair=c("single_cell","bulk_sequencing"))
diff.dat <- topTags(de.tgw,n=nrow(cds$counts))
res <- diff.dat$table
file <- paste("analysis/",sample_id,"/bulkRNAseq/DEG-Ectopic-SCvsBS_20220106.csv",sep="")
#file2 <- paste("analysis/",sample_id,"/bulkRNAseq/DEG-Top50-SCvsBS.csv",sep="")
write.csv(res[res$FDR < 0.001,],file=file)
#write.csv((row.names(res))[0:49],file=file2)

---

#### DEG on Ectopic Ovary vs sc-pseudo-bulk data
**Create Ectopic Ovary Matrix**

Ovary <- metadata[metadata$sample_type %in% c("EcO"),]
Ovary[,"X"] <- sapply(Ovary[,"X"], as.character)

sc_Ect <- sc_matrix[,(colnames(sc_matrix)) %in% c("index", Ovary$X)]
bs_Ect <- bs_matrix[,(colnames(bs_matrix)) %in% c("gene_name", Ovary$X)]
ovr_matrix <- merge(x = sc_Ect, y = bs_Ect, by.x = "index", by.y="gene_name")

dim(ovr_matrix)

**Run EdgeR**

groups <- factor(Ovary$method)
cds <- DGEList(counts = as.matrix(ovr_matrix[(2:11)]), group = groups,genes = ovr_matrix$index)
nc <- cpm(cds, normalized.lib.sizes=FALSE)
cds <- estimateCommonDisp(cds)
cds <- estimateTagwiseDisp(cds)
de.tgw <- exactTest(cds, pair=c("single_cell","bulk_sequencing"))
diff.dat <- topTags(de.tgw,n=nrow(cds$counts))
res <- diff.dat$table
file <- paste("analysis/",sample_id,"/bulkRNAseq/DEG-Ovary-SCvsBS_20220106.csv",sep="")
#file2 <- paste("analysis/",sample_id,"/bulkRNAseq/DEG-Top50-SCvsBS.csv",sep="")
write.csv(res[res$FDR < 0.001,],file=file)
#write.csv((row.names(res))[0:49],file=file2)

---
